add new game

+++ client/src/actions/game.js
@@ -1,7 +1,7 @@
 import axios from 'axios';
 import { setAlert } from './alert';
 
-import { GET_GAMES, GAME_ERROR, DELETE_GAME } from './types';
+import { GET_GAMES, ADD_GAME, DELETE_GAME, GAME_ERROR } from './types';
 
 // Get games
 export const getGames = () => async (dispatch) => {
@@ -13,19 +13,53 @@ export const getGames = () => async (dispatch) => {
   }
 };
 
-// Delete game
-export const deleteGame = (id) => async (dispatch) => {
+// Add game
+export const createGame = (formData, template, history, edit = false) => async (dispatch) => {
+  const config = {
+    headers: {
+      'Content-Type': 'application/json',
+    },
+  };
+
+  const body = {
+    ...formData,
+    template,
+  };
+
   try {
-    await axios.delete(`/api/games/${id}`);
+    const res = await axios.post('/api/games', body, config);
     dispatch({
-      type: DELETE_GAME,
-      payload: id,
+      type: ADD_GAME,
+      payload: res.data,
     });
-    dispatch(setAlert('Hra smazána.', 'success'));
+    dispatch(setAlert('Hra byla přidána.', 'success'));
+
+    if (!edit) {
+      history.push('/admin/games');
+    }
   } catch (err) {
     dispatch({
       type: GAME_ERROR,
       payload: { msg: err.response.statusText, status: err.response.status },
     });
+  }
+};
+
+// Delete game
+export const deleteGame = (id) => async (dispatch) => {
+  if (window.confirm('Opravdu chceš smazat tuto hru? Už to nejde vrátit.')) {
+    try {
+      await axios.delete(`/api/games/${id}`);
+      dispatch({
+        type: DELETE_GAME,
+        payload: id,
+      });
+      dispatch(setAlert('Hra smazána.', 'success'));
+    } catch (err) {
+      dispatch({
+        type: GAME_ERROR,
+        payload: { msg: err.response.statusText, status: err.response.status },
+      });
+    }
   }
 };
